requirements.txt
openai-whisper
faster-whisper
transformers
gradio==3.37.0
pytube
moviepy
pyyaml